requirements.txt
psycopg2-binary==2.9.10
opencv-python==4.10.0.84  
opencv-python-headless==4.10.0.84  
orjson==3.9.0 
python-dotenv==1.0.1
fastapi==0.115.5
